Logically split up client queries that wait for completion using wait() and subscriptions that wait for completion using monitor().
This implementation has separate routing and handling for all of the
various subscription types, which is an alternative to tracking
additional state on query handling to determine what is being waited
on vs monitored.

Add a subscribe_address API call (variant) that directly takes a
payment address (in addition to keeping the existing one that takes an
address hash).

Add unsubscribe_address and unsubscribe_stealth to allow server to
free up state and to cancel existing local monitoring state.